扫描Controller所在包路径下的所有Controller，作为预加载Form属性
